Code Review 1.  replace viewModel Movie to the prop in Movie 2. replace the viewModel's implement 3. add ctor for NEW and Edit

+++ ExamVidly/Models/Movie.cs
@@ -32,7 +32,7 @@ namespace ExamVidly.Models
         [Required]
         [Range(1, 20)]
         [Display(Name = "Number In Stock")]
-        public int NumberInStock { get; set; }
+        public byte NumberInStock { get; set; }
 
     }
 }
